quality of life updates for the node adapter (#9582)
* descriptive names for files and functions

* update tests

* add changeset

* appease linter

* Apply suggestions from code review


* `server-entrypoint.js` -> `server.js`

* prevent crash on stream error (from PR 9533)

* Apply suggestions from code review


* `127.0.0.1` -> `localhost`

* add changeset for fryuni's fix

* Apply suggestions from code review

* Apply suggestions from code review


---------
