+++ .changeset/unlucky-stingrays-clean.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/node": patch
+---
+
+Fixes an issue where the preview server appeared to be ready to serve requests before binding to a port.
