+++ .changeset/weak-apes-add.md
@@ -1,0 +1,6 @@
+---
+"@astrojs/node": major
+---
+
+**Breaking**: Minimum required Astro version is now 4.2.0.
+Reorganizes internals to be more maintainable.
