+++ packages/integrations/node/src/index.ts
@@ -1,6 +1,7 @@
-import type { AstroAdapter, AstroIntegration } from 'astro';
 import { AstroError } from 'astro/errors';
+import type { AstroAdapter, AstroIntegration } from 'astro';
 import type { Options, UserOptions } from './types.js';
+
 export function getAdapter(options: Options): AstroAdapter {
 	return {
 		name: '@astrojs/node',
