+++ packages/integrations/node/test/test-utils.js
@@ -2,6 +2,8 @@ import httpMocks from 'node-mocks-http';
 import { EventEmitter } from 'node:events';
 import { loadFixture as baseLoadFixture } from '../../../astro/test/test-utils.js';
 
+process.env.ASTRO_NODE_AUTOSTART = "disabled";
+process.env.ASTRO_NODE_LOGGING = "disabled";
 /**
  * @typedef {import('../../../astro/test/test-utils').Fixture} Fixture
  */
